Obstacle.h/cpp - Contains all obstacle-related classes,
Game.h/cpp - Contains the main game class and logic,

main.cpp - Just the entry point...
